Refactor pcapng: move EnhancedPacket to its own file
